ISSUE #4739 - rename helper file. delete unused file
